+++ src/components/agents/AgentFileBrowser.vue
@@ -1,14 +1,5 @@
 <template>
-  <div v-if="agent.language === 'ironpython'">
-    <v-alert prominent type="error">
-      <v-row align="center">
-        <v-col class="grow">
-          The File Browser is not yet implemented for this agent language.
-        </v-col>
-      </v-row>
-    </v-alert>
-  </div>
-  <div v-else>
+  <div>
     <execute-module-dialog
       v-model="executeDialog"
       :agent="agent.session_id"